Finished glm-ie support (largely). Also: * fixed error checks in flt_clean_windows * fixed single-precision conversion in flt_resample * added Rice Wavelet Toolbox

+++ code/filters/flt_resample.m
@@ -56,7 +56,7 @@ function [signal,state] = flt_resample(varargin)
 %                                Christian Kothe, Swartz Center for Computational Neuroscience, UCSD
 %                                2010-03-28
 
-% flt_resample_version<1.01> -- for the cache
+% flt_resample_version<1.02> -- for the cache
 
 if ~exp_beginfun('filter') return; end
 
@@ -102,7 +102,7 @@ if srate ~= signal.srate
             state.(f{1}).conds = []; end
         signal.(f{1}) = single(signal.(f{1}))';
         [tmp,state.(f{1}).conds] = upfirdn2(signal.(f{1}),state.H,state.p,state.q,state.(f{1}).conds);
-        signal.(f{1}) = tmp';            
+        signal.(f{1}) = double(tmp');            
     end
     
     % update signal meta-data
